- Reorganize the reports under a namespace - Switch registration prefix for reports to kanopi-report-
